requirements.txt
Flask==2.3.3
openpyxl==3.1.2
Pillow==10.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0

